+++ src/index.ts
@@ -7,7 +7,7 @@ import { NanoSQLDefaultBackend } from "./database/index";
 import { _NanoSQLHistoryPlugin } from "./history-plugin";
 import { NanoSQLStorageAdapter } from "./database/storage";
 
-const VERSION = 1.47;
+const VERSION = 1.48;
 
 // uglifyJS fix
 const str = ["_util"];