+++ src/utilities.ts
@@ -257,6 +257,51 @@ export const isObject = (val: any): boolean => {
 };
 
 /**
+ * Calculate Lavenshtein distance between two strings.
+ *
+ * Stolen from https://gist.github.com/andrei-m/982927
+ *
+ * @param {string} a
+ * @param {string} b
+ * @returns {number}
+ */
+export const levenshtein = (a: string, b: string): number => {
+    if (a.length === 0) return b.length;
+    if (b.length === 0) return a.length;
+    let tmp, i, j, prev, val, row;
+    // swap to save some memory O(min(a,b)) instead of O(a)
+    if (a.length > b.length) {
+        tmp = a;
+        a = b;
+        b = tmp;
+    }
+
+    row = Array(a.length + 1);
+    // init the row
+    for (i = 0; i <= a.length; i++) {
+        row[i] = i;
+    }
+
+    // fill in the rest
+    for (i = 1; i <= b.length; i++) {
+        prev = i;
+        for (j = 1; j <= a.length; j++) {
+            if (b[i - 1] === a[j - 1]) {
+                val = row[j - 1]; // match
+            } else {
+                val = Math.min(row[j - 1] + 1, // substitution
+                    Math.min(prev + 1,     // insertion
+                        row[j] + 1));  // deletion
+            }
+            row[j - 1] = prev;
+            prev = val;
+        }
+        row[a.length] = prev;
+    }
+    return row[a.length];
+};
+
+/**
  * Cast a javascript variable to a given type. Supports typescript primitives and more specific types.
  *
  * @param {string} type